大作业/backend/requirements.txt
yara-python==4.5.1
pyyaml==6.0.2
python-dateutil==2.9.0.post0
python-evtx==0.7.4
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
